Move the fields to sub-class
